change `payerAmount` type to `Decimal?`

+++ swift-vortex/Sources/Services/GetOperationDetailService/ResponseMapper/ResponseMapper+mapGetOperationDetailResponse.swift
@@ -83,7 +83,7 @@ private extension ResponseMapper {
         let payerAccountId: Int?
         let payerAccountNumber: String?
         let payerAddress: String?
-        let payerAmount: Int?
+        let payerAmount: Decimal?
         let payerCardId: Int?
         let payerCardNumber: String?
         let payerCurrency: String?
